third_party/WebKit/LayoutTests/media/track/track-load-error-readyState.html

<!DOCTYPE html>
<title>Tests the error event on HTMLTrackElement and ERROR readyState on TextTrack.</title>
<video>
    <track src="junk" default>
</video>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
async_test(function(t) {
    var track = document.querySelector("track");
    track.onerror = t.step_func_done(function() {
        assert_equals(track.readyState, HTMLTrackElement.ERROR);
    });
});
</script>
